+++ cameras/IsoCamMotionblur_t.cpp
@@ -40,9 +40,9 @@ void IsoCamMotionblur_t::raytrace(const Scene_t* scene) {
     image_->update();
 
     #pragma omp parallel for schedule(guided)
-    for (size_t index = 0; index < image_->size_y_ * image_->size_x_; ++index){
-        const size_t i = index%image_->size_y_;
-        const size_t j = index/image_->size_y_;
+    for (unsigned int index = 0; index < image_->size_y_ * image_->size_x_; ++index){
+        const unsigned int i = index%image_->size_y_;
+        const unsigned int j = index/image_->size_y_;
         Vec3f col = Vec3f(); // Or declare above?
         for (unsigned int k = 0; k < subpix_[0]; k++){
             for (unsigned int l = 0; l < subpix_[1]; l++){
